Restructure old file placement for easier organization
