+++ src/scripts/legacy/globals.js
